page和worker代码优化
