+++ .github/workflows/cloudflare-worker-deploy.yml
@@ -12,6 +12,13 @@ jobs:
     timeout-minutes: 60
     steps:
       - uses: actions/checkout@v2
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v4.0.0
+        with:
+          run_install: true
+          version: 9
+      - name: build - 构建
+        run: pnpm run build-worker
       - name: Build & Deploy Worker
         uses: cloudflare/wrangler-action@v3
         with:
