+++ cloudflare/rollup.config.workers.mjs
@@ -1,0 +1,20 @@
+import compiler from '@ampproject/rollup-plugin-closure-compiler';
+
+// rollup.config.mjs
+// ---cut-start---
+/** @type {import('rollup').RollupOptions} */
+// ---cut-end---
+export default {
+    input: 'cloudflare/worker.js',
+    output: [
+        {
+            file: 'worker.js',
+            format: 'es',
+        }
+    ],
+    plugins:[
+        compiler({
+            compilation_level:"ADVANCED"
+        })
+    ]
+};
