+++ package.json
@@ -1,0 +1,14 @@
+{
+  "name": "microsoft-copilot-porxy",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "build-worker": "rollup --config ./cloudflare/rollup.config.workers.mjs",
+    "build-page": "rollup --config ./cloudflare/rollup.config.pages.mjs"
+  },
+  "devDependencies": {
+    "@ampproject/rollup-plugin-closure-compiler": "^0.27.0",
+    "rollup": "^4.18.0",
+    "wrangler": "^3.0.0"
+  }
+}
