+++ wrangler.toml
@@ -1,5 +1,5 @@
 name = "go-proxy-bingai"
-main = "cloudflare/worker.js"
+main = "worker.js"
 compatibility_date = "2024-04-29"
 workers_dev = true
 rules = [
